\name{fmegen}
\alias{fmegen}
\alias{chr}
\alias{posi}
\alias{gen}
\title{
 Genotype data 
}
\description{
Numeric format of genotype dataset.
}
\usage{
  data(fmegen)
}
\details{
Dataset input of Genotype for FASTmrEMMA function. 
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}
