\name{fmegenraw}
\alias{fmegenraw}
\alias{fmers#}
\alias{fmechrom}
\alias{fmepos}
\title{
 raw genotype data 
}
\description{
Numeric format of raw genotype dataset.
}
\usage{
  data(fmegenraw)
}
\details{
Dataset input of raw genotype for FASTmrEMMA function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
