\name{mrgen}
\alias{mrgen}
\alias{chromosome}
\alias{position}
\alias{genotype}
\title{
 Genotype data 
}
\description{
Numeric format of genotype dataset.
}
\usage{
  data(mrgen)
}
\details{
Dataset input of Genotype for mrMLM function. 
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}