\name{mrgenraw}
\alias{mrgenraw}
\alias{rs#}
\alias{chrom}
\alias{pos}
\title{
 raw genotype data 
}
\description{
Numeric format of raw genotype dataset.
}
\usage{
  data(mrgenraw)
}
\details{
Dataset input of raw genotype for mrMLM function. 
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}